refactor: Moving auth module the new architecture
