+++ src/modules/Auth/views/Login/components/__tests__/formLoginInteractingWith.spec.ts
@@ -3,7 +3,7 @@ import { mount } from '@vue/test-utils'
 import FormLogin from '../FormLogin.vue'
 import { AppButton, AppInput, AppErrorMessage, AppFlashMessage, AppSelect } from '@/modules/__tests__/globalComponents'
 import type { VueWrapper } from '@vue/test-utils'
-import type { UpdateUser } from '../../../../Auth/types/Auth'
+import type { UpdateUser } from '@/modules/Auth/types/Auth'
 
 type Wrapper = VueWrapper<Partial<{
   email: string;
