+++ src/modules/Auth/views/Profile/components/ProfileFileUpload.vue
