+++ src/modules/Auth/views/Profile/components/ProfileUser/__tests__/spyProfileUpdateUser.spec.ts
@@ -1,20 +1,24 @@
 import { describe, it, vi, expect } from 'vitest'
 import { mount } from '@vue/test-utils'
 import { createTestingPinia } from '@pinia/testing'
-import ProfileUser from '../ProfileUser.vue'
-import { updateUser } from "@/modules/Auth/services"
-import * as MockService from "@/modules/__tests__/apiMock/AuthService"
+import ProfileUser from '../Index.vue'
 import { AppButton, AppInput, AppErrorMessage, AppFlashMessage, AppSelect } from '@/modules/__tests__/globalComponents'
-import type { Mock } from 'vitest'
+import type { VueWrapper } from '@vue/test-utils'
+import type { UpdateUser } from '@/modules/Auth/types/Auth'
 
 const payload = {
   name: "John Doe",
   email: "user@email.ext"
 }
 
-describe('ModuleauthComponentProfileUser.vue',  async () => {
+describe('ModuleAuthComponentProfileUser.vue',  () => {
+
+  type Wrapper = VueWrapper<Partial<{
+    updateUser: (user: UpdateUser) => void
+  }>> // Or put "defineExpose({ updateUser })" in components
+
   it('should be called updateUser with payload', async () => {
-    const wrapper = mount(ProfileUser, {
+    const wrapper: Wrapper = mount(ProfileUser, {
       global: {
         components: {
           AppButton,
@@ -30,11 +34,12 @@ describe('ModuleauthComponentProfileUser.vue',  async () => {
         })]
       }
     })
+    const updateUserSpy = vi.spyOn(wrapper.vm, 'updateUser');
 
-    vi.mock("@/modules/Auth/services");
-    (updateUser as Mock).mockImplementation(MockService.updateUser); 
-    
-    expect(await updateUser({})).toBe("User updated.");
+    wrapper.vm.updateUser(payload)
 
+    expect(updateUserSpy).toHaveBeenCalled()
+    expect(updateUserSpy).toHaveBeenCalledTimes(1)
+    expect(updateUserSpy).toHaveBeenCalledWith(payload)
   })  
 })
