+++ src/modules/Auth/views/Register/components/FormRegister.vue
