+++ src/modules/Auth/views/Register/components/__tests__/formRegisterMountedCorrectly.spec.ts
