+++ src/modules/Auth/views/Register/composables/useRegister.ts
