+++ src/modules/Authorization/views/Menus/Datagrid/Index.vue
@@ -4,8 +4,8 @@ import { defineComponent } from 'vue'
 import * as MenuService from "@/modules/Authorization/services/MenuService"
 import AppPaginationB from "@/core/components/AppPaginationB.vue";
 import AppPageHeader from "@/core/components/AppPageHeader.vue"
-import Create from './Create.vue'
-import Edit from './Edit.vue'
+import Create from './components/Create.vue'
+import Edit from './components/Edit.vue'
 
 export default defineComponent({  
   metaInfo: { title: 'Menus' },
